requirements.txt
# Thư viện cần thiết
opencv-python
torch
yolov5
pyyyaml
